streamlit_ui/requirements.txt
streamlit==1.37.1
requests==2.32.3
Pillow>=10.0.0
